Pull `Parser` interface to the package level
